pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "varro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "httpx>=0.28.1",
    "dill>=0.4.0",
    "jinja2>=3.1.6",
    "matplotlib>=3.10.7",
    "networkx>=3.5",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "psycopg[binary]>=3.2.12",
    "pyarrow>=22.0.0",
    "pydantic-ai>=1.7.0",
    "pytest-playwright>=0.7.2",
    "rapidfuzz>=3.14.3",
    "sqlmodel>=0.0.27",
    "tabulate>=0.9.0",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "python-fasthtml>=0.12.37",
    "ipython>=9.6.0",
    "mistletoe>=1.5.1",
    "resend>=2.19.0",
    "msgpack>=1.1.2",
    "zstandard>=0.25.0",
    "logfire>=4.14.2",
    "safecmd>=0.1.1",
    "geopandas>=1.1.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]


[tool.hatch.build.targets.wheel]
packages = ["varro", "ui"] 
